@@ -1,3 +1,15 @@
-module.exports = {
-  presets: ["@vue/app"]
-};
+/* eslint-env node */
+
+module.exports = api => {
+  return {
+    presets: [
+      [
+        '@quasar/babel-preset-app',
+        api.caller(caller => caller && caller.target === 'node')
+          ? { targets: { node: 'current' } }
+          : {}
+      ]
+    ]
+  }
+}
+
