@@ -1,8 +1,8 @@
 <template>
   <q-select
     outlined
-    :value="value"
-    v-on:input="$emit('input', $event)"
+    :model-value="value"
+    @update:model-value="$emit('input', $event)"
     :options="options"
     :option-label="item => (typeof item === 'object' ? $tr(item.label) : item)"
     :label="$tr('general.form.fields.country')"
@@ -18,9 +18,10 @@
 </template>
 
 <script>
+/* eslint-disable */
 export default {
-  name: "CountrySelect",
-  props: ["value"],
+  name: 'CountrySelect',
+  props: ['value'],
   data() {
     return {
       options: [],
@@ -28,22 +29,20 @@ export default {
 
       // Default countries based on locale
       defaultCountries: {
-        cs: "CZ"
-      }
+        cs: 'CZ',
+      },
     };
   },
   created() {
-    this.loadCountries().then(data => {
+    this.loadCountries().then((data) => {
       this.allOptions = data;
 
-      let value = this.value;
+      let { value } = this;
 
       // Pick default option based on selected locale
-      if (this.defaultCountries[this.$i18n.locale] && !value)
-        value = this.defaultCountries[this.$i18n.locale];
+      if (this.defaultCountries[this.$i18n.locale] && !value) value = this.defaultCountries[this.$i18n.locale];
 
-      if (value && typeof value === "string")
-        this.$emit("input", this.getCountryByCode(value));
+      if (value && typeof value === 'string') this.$emit('input', this.getCountryByCode(value));
 
       this.options = data;
     });
@@ -51,25 +50,24 @@ export default {
   methods: {
     loadCountries() {
       return new Promise((resolve, reject) => {
-        if (this.$db("countries-select"))
-          return resolve(this.$db("countries-select"));
+        if (this.$db('countries-select')) return resolve(this.$db('countries-select'));
 
         this.$api({
-          url: "country",
-          method: "get"
+          url: 'country',
+          method: 'get',
         })
-          .then(d => {
-            let countries = [];
-            let data = d.data;
+          .then((d) => {
+            const countries = [];
+            const { data } = d;
 
-            for (let code of Object.keys(data)) {
+            for (const code of Object.keys(data)) {
               countries.push({
                 value: code,
-                label: data[code]
+                label: data[code],
               });
             }
 
-            this.$db("countries-select", countries);
+            this.$db('countries-select', countries);
             resolve(countries);
           })
           .catch(reject);
@@ -77,7 +75,7 @@ export default {
     },
 
     getCountryByCode(code) {
-      let filtered = this.allOptions.filter(item => item.value === code);
+      const filtered = this.allOptions.filter((item) => item.value === code);
 
       if (filtered.length) return filtered[0];
 
@@ -86,7 +84,7 @@ export default {
 
     filterFn(val, update) {
       val = val.trim();
-      if (val === "") {
+      if (val === '') {
         update(() => {
           this.options = this.allOptions;
         });
@@ -95,17 +93,18 @@ export default {
 
       update(() => {
         const needle = val.toLowerCase();
-        this.options = this.allOptions.filter(item => {
-          if (typeof item === "object")
+        this.options = this.allOptions.filter((item) => {
+          if (typeof item === 'object') {
             return (
               this.$tr(item.label)
                 .toLowerCase()
                 .includes(needle) || item.value.toLowerCase().includes(needle)
             );
+          }
           return false;
         });
       });
-    }
-  }
+    },
+  },
 };
 </script>
