@@ -1,7 +1,7 @@
 <template>
   <div>
     <q-card flat bordered class="my-card bg-grey-1 q-mb-md">
-      <q-card-section class="header" @click="$emit('toggleVisibility', id)">
+      <q-card-section class="header" @click="toggleCard">
         <div class="row items-center no-wrap">
           <div class="col">
             <q-btn
@@ -39,31 +39,37 @@
           </div>
         </div>
       </q-card-section>
-      <slide-up-down :active="visible" :duration="500">
-        <q-card-section>
+      <transition-group
+          appear
+          enter-active-class="animated slideInDown"
+          leave-active-class="animated slideOutUp"
+      >
+        <q-card-section v-if="showCard">
           <form-fields
-            ref="form-fields"
-            @input="catchInput"
-            :autofill="autofill"
-            :is-team="true"
-            :accommodationType="accommodationType"
-            :mealType="mealType"
-            :possibleDiets="possibleDiets"
-            :role="1"
-            :requireEmail="requireEmail"
+              ref="form-fields"
+              @input="catchInput"
+              :autofill="autofill"
+              :is-team="true"
+              :accommodationType="accommodationType"
+              :mealType="mealType"
+              :possibleDiets="possibleDiets"
+              :role="1"
+              :requireEmail="requireEmail"
           />
         </q-card-section>
-      </slide-up-down>
+      </transition-group>
     </q-card>
   </div>
 </template>
 
 <script>
-import formFields from "./FormFields";
+/* eslint-disable */
+import formFields from './FormFields';
+
 export default {
-  name: "TeamPersonCard",
+  name: 'TeamPersonCard',
   components: {
-    formFields
+    formFields,
   },
   props: {
     autofill: Object,
@@ -74,19 +80,24 @@ export default {
     accommodationType: String,
     mealType: String,
     possibleDiets: Array,
-    requireEmail: Boolean
+    requireEmail: Boolean,
   },
   data() {
     return {
-      translationPrefix: "tournament.",
-      formData: {}
+      showCard: true,
+      translationPrefix: 'tournament.',
+      formData: {},
     };
   },
   methods: {
     catchInput(data) {
       this.formData = data;
-      this.$emit("input", data, this.id);
+      this.$emit('input', data, this.id);
+    },
+    toggleCard() {
+      this.showCard = ! this.showCard;
+      this.$emit('toggleVisibility', this.id);
     }
-  }
+  },
 };
 </script>
