@@ -1,7 +1,7 @@
 <template>
   <q-layout
     view="lHh Lpr lFf"
-    :class="'bg-grey-2 page-' + $router.currentRoute.name"
+    :class="'bg-grey-2 page-' + $route.name"
   >
     <q-header elevated>
       <q-toolbar>
@@ -20,12 +20,12 @@
               <img src="../assets/logo.svg" alt="logo" />
             </q-avatar>
             greybox 2.0
-            <template v-if="env.VUE_APP_STAGE === 'pds'">PDS</template>
-            <span v-if="env.VUE_APP_MODE !== 'production'" class="mode-flag">
-              <template v-if="env.VUE_APP_STAGE === 'debug'">
+            <template v-if="env.STAGE === 'pds'">PDS</template>
+            <span v-if="env.MODE !== 'production'" class="mode-flag">
+              <template v-if="env.STAGE === 'debug'">
                 debug
               </template>
-              <template v-else-if="env.VUE_APP_STAGE === 'local'">
+              <template v-else-if="env.STAGE === 'local'">
                 dev
               </template>
               <template v-else>
@@ -139,23 +139,24 @@
 </template>
 
 <script>
-import Sidenav from "./components/Sidenav";
-import { EventBus } from "../event-bus";
+/* eslint-disable */
+import Sidenav from './components/Sidenav';
+import { EventBus } from '../event-bus';
 
 export default {
-  name: "LayoutDefault",
+  name: 'LayoutDefault',
 
   components: {
-    Sidenav
+    Sidenav,
   },
 
   data() {
     return {
       leftDrawerOpen:
-        this.$q.platform.is.desktop &&
-        localStorage.getItem("leftDrawerOpen") !== "false",
+        this.$q.platform.is.desktop
+        && localStorage.getItem('leftDrawerOpen') !== 'false',
       user: null,
-      fullLoader: 0 // number of active loadings
+      fullLoader: 0, // number of active loadings
     };
   },
 
@@ -163,7 +164,7 @@ export default {
     // Show loading until events load
     this.fullLoader = 1;
 
-    EventBus.$on("fullLoader", value => {
+    this.$bus.$on('fullLoader', (value) => {
       if (value) this.fullLoader++;
       else this.fullLoader--;
 
@@ -174,35 +175,34 @@ export default {
   methods: {
     toggleDrawerMenu() {
       this.leftDrawerOpen = !this.leftDrawerOpen;
-      localStorage.setItem("leftDrawerOpen", this.leftDrawerOpen);
+      localStorage.setItem('leftDrawerOpen', this.leftDrawerOpen);
     },
     switchLocale(locale) {
       if (this.$i18n.locale !== locale) {
-        //current URL
-        let originalPath = this.$tr(
-          "paths." + this.$router.resolve({}).route.name
+        // current URL
+        const originalPath = this.$tr(
+          `paths.${this.$route.name}`,
         );
 
-        //change locale
+        // change locale
         this.$i18n.locale = locale;
 
-        //new URL
-        let newPath = this.$tr("paths." + this.$router.resolve({}).route.name);
+        // new URL
+        const newPath = this.$tr(`paths.${this.$route.name}`);
 
-        //get URL from router
-        let url = this.$router.resolve({});
-        url = url.location;
+        // get URL from router
+        let url = this.$route;
 
         // Homepage cases
-        if (originalPath == "") url.path = "/en/";
-        else if (newPath == "") url.path = "/";
+        if (originalPath === '') url.path = '/en/';
+        else if (newPath === '') url.path = '/';
         // replace url in router with localized one
         else url.path = url.path.replace(originalPath, newPath);
 
         // go to new url
         this.$router.push(url);
       }
-    }
-  }
+    },
+  },
 };
 </script>
