@@ -10,6 +10,6 @@
 
 <script>
 export default {
-  name: "PageAbout"
+  name: 'PageAbout',
 };
 </script>
