@@ -66,17 +66,18 @@
 </template>
 
 <script>
+/* eslint-disable */
 export default {
-  name: "PageNewPassword",
+  name: 'PageNewPassword',
   data() {
     return {
-      translationPrefix: "auth.",
+      translationPrefix: 'auth.',
       password: null,
       passwordConfirmation: null,
       isPwd: true,
       isPwd2: true,
       loading: false,
-      token: null
+      token: null,
     };
   },
   methods: {
@@ -84,64 +85,65 @@ export default {
       this.loading = true;
 
       this.$api({
-        url: "reset",
+        url: 'reset',
         sendToken: false,
         data: {
           password: this.password,
           password_confirmation: this.passwordConfirmation,
-          token: this.token
+          token: this.token,
         },
         alerts: false,
-        method: "put"
+        method: 'put',
       })
         .then(() => {
-          this.$flash(this.$tr("passwordReset.successReset"), "success");
-          this.$router.replace(this.$path("auth.login"));
+          this.$flash(this.$tr('passwordReset.successReset'), 'success');
+          this.$router.replace(this.$path('auth.login'));
         })
-        .catch(data => {
+        .catch((data) => {
           if (data.response.data) {
-            let response = data.response.data;
+            const response = data.response.data;
 
             // Just one error
             if (response.message) {
               this.$flash(
-                this.$tr("passwordReset.validation." + response.message),
-                "error",
+                this.$tr(`passwordReset.validation.${response.message}`),
+                'error',
                 false,
-                9000
+                9000,
               );
 
               // Invalid token -> redirect to generate a new one
-              if (response.message === "tokenNotFound")
-                this.$router.replace(this.$path("auth.passwordReset"));
+              if (response.message === 'tokenNotFound') this.$router.replace(this.$path('auth.passwordReset'));
             } // More errors - same as in sign up
-            else
-              for (let index in data.response.data)
-                data.response.data[index].forEach(message => {
+            else {
+              for (const index in data.response.data) {
+                data.response.data[index].forEach((message) => {
                   this.$flash(
                     this.$tr(
-                      "signUp.validation." +
-                        index +
-                        "." +
-                        message.substr(11).replace(".", "-")
+                      `signUp.validation.${
+                        index
+                      }.${
+                        message.substr(11).replace('.', '-')}`,
                     ),
-                    "error",
+                    'error',
                     false,
-                    9000
+                    9000,
                   );
                 });
-          } else this.$flash(this.$tr("general.error", null, false), "error");
+              }
+            }
+          } else this.$flash(this.$tr('general.error', null, false), 'error');
         })
         .finally(() => {
           this.loading = false;
         });
-    }
+    },
   },
   created() {
     this.token = this.$route.params.token;
 
-    if (this.$auth.check()) this.$router.replace({ name: "home" });
-  }
+    if (this.$auth.check()) this.$router.replace({ name: 'home' });
+  },
 };
 </script>
 
