@@ -28,6 +28,6 @@
 
 <script>
 export default {
-  name: "Homepage"
+  name: 'Homepage',
 };
 </script>
